refactor(21): extract flattenZod to shared admin module (CLEAN-03)
The flattenZod ZodError->field-map helper was copy-pasted in all 6 admin
actions.ts. Extracted to src/lib/admin/zod-errors.ts; the 6 files import it.
ActionResult left per-file (only 3 define it and blog's id is required,
not optional, so sharing would wrongly widen the type). NewsletterSignup
self-duplication kept by decision (two independently-styled variant render
branches; deduping would couple them + risk visual regressions with no
visual-regression test).

[hudsor01]
