+++ src/app/(admin)/admin/blog/actions.ts
@@ -18,7 +18,6 @@
  */
 import { revalidatePath } from 'next/cache'
 import { redirect } from 'next/navigation'
-import type { ZodError } from 'zod'
 import { requireAdminSession } from '@/lib/admin/auth'
 import {
 	createBlogPost,
@@ -28,6 +27,7 @@ import {
 } from '@/lib/admin/blog-queries'
 import { isUniqueViolation } from '@/lib/admin/db-errors'
 import { formDataToObject } from '@/lib/admin/form-data'
+import { flattenZod } from '@/lib/admin/zod-errors'
 import { logger } from '@/lib/logger'
 import {
 	createAdminBlogPostSchema,
@@ -37,17 +37,6 @@ import {
 export type ActionResult =
 	| { ok: true; id: string }
 	| { ok: false; errors: Record<string, string> }
-
-function flattenZod(error: ZodError): Record<string, string> {
-	const out: Record<string, string> = {}
-	for (const issue of error.issues) {
-		const key = issue.path.length > 0 ? issue.path.join('.') : '_form'
-		if (!(key in out)) {
-			out[key] = issue.message
-		}
-	}
-	return out
-}
 
 export async function createBlogPostAction(
 	formData: FormData
